refactor(daemon): wire Transport/DnsService/ConnectionManager service bodies
Content of the service extraction (the prior commit only carried the file
rename). Strangler phase 1: build each service alongside MeshManager, tree stays
green (build + clippy + 357 tests).

- Transport (foundation.rs) constructed in bootstrap, exposed via transport().
- DnsService: Arc-wrapped, sync_network/clear_network/resolve added; two call
  sites migrated off direct table access.
- ConnectionManager: owns the per-peer driver + demux + handler registry +
  pending_pongs; ProtocolRouter slimmed to a thin ALPN dispatcher with
  passthroughs (register/is_registered/unregister/set_mesh_dispatch/pending_pongs/
  drive_mesh_connection).
